decided to keep this in the main file, as it involves ros msg translation
