+++ src/perception/lidar_pipeline/lidar_pipeline/node_detector.py
@@ -9,9 +9,51 @@ from rclpy.subscription import Subscription
 
 from driverless_msgs.msg import Cone, ConeDetectionStamped
 from geometry_msgs.msg import Point
-from sensor_msgs.msg import PointCloud2
+from sensor_msgs.msg import PointCloud2, PointField
 
 from .process import *
+
+
+def fields_to_dtype(fields, point_step):
+    """
+    FROM ROS2_NUMPY
+    Convert a list of PointFields to a numpy record datatype.
+    """
+    DUMMY_FIELD_PREFIX = "__"
+    # mappings between PointField types and numpy types
+    type_mappings = [
+        (PointField.INT8, np.dtype("int8")),
+        (PointField.UINT8, np.dtype("uint8")),
+        (PointField.INT16, np.dtype("int16")),
+        (PointField.UINT16, np.dtype("uint16")),
+        (PointField.INT32, np.dtype("int32")),
+        (PointField.UINT32, np.dtype("uint32")),
+        (PointField.FLOAT32, np.dtype("float32")),
+        (PointField.FLOAT64, np.dtype("float64")),
+    ]
+    pftype_to_nptype = dict(type_mappings)
+
+    offset = 0
+    np_dtype_list = []
+    for f in fields:
+        while offset < f.offset:
+            # might be extra padding between fields
+            np_dtype_list.append(("%s%d" % (DUMMY_FIELD_PREFIX, offset), np.uint8))
+            offset += 1
+
+        dtype = pftype_to_nptype[f.datatype]
+        if f.count != 1:
+            dtype = np.dtype((dtype, f.count))
+
+        np_dtype_list.append((f.name, dtype))
+        offset += pftype_to_nptype[f.datatype].itemsize * f.count
+
+    # might be extra padding between points
+    while offset < point_step:
+        np_dtype_list.append(("%s%d" % (DUMMY_FIELD_PREFIX, offset), np.uint8))
+        offset += 1
+
+    return np_dtype_list
 
 
 def cone_msg(x: float, y: float) -> Cone:
